every entry its own file in the models subdir. made a utilfunction moveProductInputFieldBelowMe() for needsScreen
