+++ src/lib/components/header-input-field-list-tile.dart
@@ -4,12 +4,8 @@ import 'dart:convert';
 import 'dart:math';
 import 'package:path_provider/path_provider.dart';
 import 'package:flutter/material.dart';
-import 'package:uuid/uuid.dart';
 
-import 'package:shoplist/models/appdata.dart';
-import 'package:shoplist/models/shop-list-entry.dart';
-import 'package:shoplist/util/fs.dart';
-import 'package:shoplist/util/misc.dart';
+import 'package:shoplist/models/header-input-field.dart';
 
 ListTile headerInputFieldListTile(
     HeaderInputField entry,
